更新PWM库

+++ _Keilduino (STM32F0xx)/Core/pwm.h
@@ -7,7 +7,7 @@ extern "C" {
 
 #include "mcu_type.h"
 
-uint8_t PWM_Init(uint8_t Pin, uint16_t PWM_DutyCycle, uint16_t PWM_Frequency);
+uint8_t PWM_Init(uint8_t Pin, uint16_t PWM_DutyCycle, uint32_t PWM_Frequency);
 uint16_t pwmWrite(uint8_t Pin, uint16_t val);
 void TIMx_Init(TIM_TypeDef* TIMx, uint16_t arr, uint16_t psc, uint8_t TimerChannel);
 uint8_t Get_TIMx_GPIO_AF_x(uint8_t Pin);
